+++ tests/dft/test_build.py
@@ -9,12 +9,8 @@ from configurations import dft
 @dft
 class TestBuild:
     def test_build_dft(self):
-        model = stormpy.dft.build_sparse_model_from_json_dft(get_example_path("dft", "and.json"))
+        dft = stormpy.dft.load_dft_json(get_example_path("dft", "and.json"))
         formulas = stormpy.parse_properties("T=? [ F \"failed\" ]")
-        assert model.nr_states == 4
-        assert model.nr_transitions == 5
-        assert len(model.initial_states) == 1
-        initial_state = model.initial_states[0]
-        assert initial_state == 1
-        result = stormpy.model_checking(model, formulas[0])
-        assert math.isclose(result.at(initial_state), 3)
+        assert dft.nr_elements() == 3
+        results = stormpy.dft.analyze_dft(dft, [formulas[0].raw_formula])
+        assert math.isclose(results[0], 3)